+++ server.js
@@ -3,7 +3,7 @@ require('env-yaml').config({ path: './.env.yaml' });
 
 var express = require("express");
 var exphbs = require("express-handlebars");
-var db = require("./models");
+var db = require("./server/models");
 
 var app = express();
 var PORT = process.env.PORT || 3000;
@@ -24,8 +24,8 @@ app.engine(
 app.set("view engine", "handlebars");
 
 // Routes
-require("./routes/apiRoutes")(app);
-require("./routes/htmlRoutes")(app); 
+require("././server/routes/apiRoutes")(app);
+require("././server/routes/htmlRoutes")(app); 
 
 console.clear();
 var syncOptions = {};
