+++ views/index.handlebars
@@ -1,6 +1,3 @@
-
-<br>
-
 {{!-- Skillsets --}}
 <div class="container-fluid col-md-9 col-11 px-0" data-aos="fade-up" data-aos-duration="2000">
     <div>
